primordial SWO/SWV support for DAPLink
This support has been tested on an LPC-Link2 acting as the debugger
for a NXP LPC55S69. Unfortunately, it doesn't quite work:  for
reasons yet unknown, the LPC-Link2 seems to not reply to any data
read after the first SWO data read.  (The first SWO data read seems
correct.)  Nonetheless, hopefully this implementation can serve as
a baseline for those wishing to implement SWO/SWV on DAPLink!