MEM-AP operations fail if SPIDEN is set
Bit 30 in the AHB-AP specified the security of the AHB transfer output
on SLVNONSEC: if the bit is set, it's a non-secure transfer; if the bit
is clear it's a secure transfer.  This bit must therefore be the inverse
of SPIDEN: if SPIDEN is set, bit 30 must be clear, and if SPIDEN is clear,
bit 30 must be set.