daplink probe gets stuck on FT4232H MiniModule on illumos
On illumos systems the USB stack does not allow control transfers to
devices that are bound to drivers and choose not to expose ugen.  The
daplink probe routine attempts to make transfers to the FTDI FT4232H
MiniModule UART board and fails.  As this is not even a candidate
device, rather than throw up our hands we should keep scanning with the
next available device.