Stop unconditionally using the debug mailbox
In debug_port_start we currently unconditionally call
enable_debug_mailbox. This will certainly always work to make sure
there is a valid access port but has the unwanted side effect of
making it so we can't attach to a running target. Per the NXP
manual, if there is a valid image in flash and/or we're attached
to a running target there should be no need to use the debug mailbox.
Just remove the call to enable the debug mailbox for now. We can
use the debug mailbox for other cases such as halting after reset.